requirements.txt
cryptography==43.0.3
Flask==3.0.3
Pillow==11.0.0
